requirements.txt
streamlit
opencv-python
tensorflow
keras
Pillow
numpy
pandas
matplotlib
scikit-learn